Check kube certs after issue in teleterm gateways
This makes sure that Connect does the same checks as tsh.
